chore: move race into a separate file
