+++ src/operators/index.ts
@@ -65,7 +65,8 @@ export { publish } from '../internal/operators/publish';
 export { publishBehavior } from '../internal/operators/publishBehavior';
 export { publishLast } from '../internal/operators/publishLast';
 export { publishReplay } from '../internal/operators/publishReplay';
-export { race, raceWith } from '../internal/operators/raceWith';
+export { race } from '../internal/operators/race';
+export { raceWith } from '../internal/operators/raceWith';
 export { reduce } from '../internal/operators/reduce';
 export { repeat } from '../internal/operators/repeat';
 export { repeatWhen } from '../internal/operators/repeatWhen';
